Use pipe instead of interceptor to parse Change
